+++ ios/Classes/RadarFlutterPlugin.m
@@ -134,9 +134,7 @@
         [self sendEvent:call withResult:result];        
     } else if ([@"getMatrix" isEqualToString:call.method]) {
         [self getMatrix:call withResult:result];        
-    } else if ([@"startForegroundService" isEqualToString:call.method]) {
-        // do nothing
-    } else if ([@"stopForegroundService" isEqualToString:call.method]) {
+    } else if ([@"setForegroundServiceOptions" isEqualToString:call.method]) {
         // do nothing
     } else {
         result(FlutterMethodNotImplemented);